haneul-http: add support for insecure connections when a tls config is provided
Add support for allowing insecure connections when a tls config is
provided.

Normally when a server is started with a TLS config, it is required that
all clients perform a TLS handshake. This patch optionally allows a
client to skip the TLS handshake by simply starting http/1.1 or
http/2 communication.

This is intended to be a temporary feature while the haneul validator
interface is being migrated to support TLS secured connections.