Implement consistency pruning in the builit-in pruner
## Description 

This PR implements consistent pruning within the built-in pruner for
concurrent pipelines.
Instead of relying on a separate pipeline to do pruning, this PR adds
shared data between the processor and pruner for obj_info.
Note that this PR isn't finished yet, sending our early to make sure
this is not in the wrong direction.
To implement this, I also need to make pruner take `self` so that it has
access to shared data. The first checkpoint of indexer also needs
special care to make sure for any pipeline that requires processed data
to prune, we must start processing at the pruner watermark instead of
commit watermark.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: