[consensus increase round prober default timeout to 4s
## Description 

It should be ok to exclude responses when they are taking >= 2s to
return, for quorum round calculations. But it seems better to use a
longer timeout to reduce variance in the computed quorum rounds among
validators.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: