[docgen] Rewrite docgen on Move model v2
## Description 

- Docs are now on the AST
- Docgen uses move-model-2
  - There is a big knot of work in the package system for this
  - The tests are their own crate due to circular dep issues
- Functionally, there is not much of a change, except that we use named
addresses in docs! Seems like a win to me but we should discuss as its a
bit confusing

## Test plan 

- Migrated tests
- Regenerated docs 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [X] CLI: Move docgen now uses named addresses. This will change some directory structures in the output format, which may break any tools targeting a rigid directory structure. 
- [ ] Rust SDK:

---------