refactor(json-rpc): avoid taking package resolver by value
## Description

The package resolver is passed in an `Arc` in case the callee needs to
issue concurrent requests (which each need their own package resolver),
but it does not need to be passed by value down to the leaves. It can be
passed by reference and cloned to get a value if that is necessary (most
of the time, it is not).

## Test plan

CI