fix(indexer-alt): generate_schema also run framework schema
## Description
When running `generate_schema`, for the indexer, it should also run the
indexing framework's migrations, but then exclude the framework's tables
when printing the schema. This mirrors the behaviour when the indexer
runs migrations, and it allows the indexer's migrations to depend on
framework tables existing.

## Test plan

A later change will introduce a migration that expects the `watermarks`
table to exist. That fails before this change (when run on a clean DB),
but now succeeds.