test-adapter: add named variable for transaction digests
## Description

For each transaction that is executed by the test adapter, record a
named variable, `digest_$task` where `$task` is the transactional
testing task number (the same number that appears in the first position
of fake object IDs).

This is to allow later tasks to refer to transaction digests as
variables in their bodies (in particular JSON RPC and GraphQL requests).

This is in preparation for adding tests for querying transaction blocks
by their digest.

The initial implementation assumes that all transactional test commands
execute at most one transaction (which is true for now). The adapter
will panic if this is ever not true, at which point we can do something
more sophisticated, if we need/want to.

## Test plan

Run existing transactional tests and make sure they all still pass
(makes sure the assumption we made about the number of transactions run
holds):

```
$ cargo nextest run            \
  -p haneul-indexer-alt-e2e-tests \
  -p haneul-graphql-e2e-tests     \
  -p haneul-adapter-transactional-tests
```

A follow-up change will introduce tests that make use of the new
feature.