CheckpointService refactoring (for data quarantining)
This change adds a replay-on-startup phase to checkpoint service
construction.

The basic idea is that with data quarantining, if the validator crashes,
it will reprocess consensus commits that had already been processed, but
whose output had not yet been committed to disk. Constructed checkpoints
are one of the outputs of this process.

When the node starts up, it should be in the same state it was in when
it crashed. In order to do this, we need to replay all the consensus
commits that had been previously processed. This will cause checkpoints
to be constructed. The effect of this change is that the node will wait
until these checkpoints have been reconstructed before entering normal
operations.