[Consensus] Disable the new linearizer logic
## Description 

This is running on PT nightly for the past couple of days and a very
small degradation in performance has been observed. It's not entirely
clear whether is due to the new linearizer logic - metrics showcase an
increase in the utilisation for the related function - but I am
disabling for now until get better clarity.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: