fix(rpc-alt): queryTransactionBlocks transaction not found
## Description
When using `response::transaction` to convert a stored transaction into
a `HaneulTransactionBlockResponse`, as part of `queryTransactionBlocks`, we
also need to detect a `NotFound` error, (usually a user error), and
translate that into an internal error, because we shouldn't be able to
query a transaction block that we can't get the contents for.

## Test plan
Tested against our production database which doesn't include
`kv_transactions` (so does trigger this internal error).