refactor(rpc-alt): extract out load_latest
## Description
Pull out the data loading part of `fetch_latest_for_system_state` to be
used in other scenarios. The shared part performs one data loader
look-up to get the latest version of an object, and then another to get
the contents at that version.

The other option was to implement this as its own data loader, but I
opted against that because while that would have shared implementation
logic, requests for latest versions and contents from this new data
loader would not have been batched with requests from the existing data
loaders.

## Test plan
CI