test(rpc-alt): Introducing FullCluster
## Description

Add a new type to the library supporting E2E tests that combines an
off-chain cluster with an instance of simulacrum. This can be used to
drive E2E tests that need to configure indexing and RPC to a greater
degree than is possible with transactional tests.

This type will be used to write tests for HaneulNS-related JSONRPC-methods,
which require us to publish some code on-chain and extract some IDs from
that process before we set-up the indexer.

## Test plan

Will be used for writing tests in a follow-up change.