rpc-alt: testing query transaction + pruning
## Description

Adding a test to make sure that transaction querying continues to work
in the presence of pruning, because there is some tricky logic related
to applying a lowerbound to the filtered results.

At the same time, I also created a testing-specific config for the
indexer, which speeds up polling but reduces concurrency -- this shaves
off about 10 to 15s from local testing runs overall, and 30s for the
newly added test (which generates a lot of checkpoints).

## Test plan

```
haneul$ cargo nextest run -p haneul-indexer-alt-e2e-tests \
 --test transaction_pruning_tests
```