rpc-alt: redefine TransactionFilter
## Description

Rather than re-using the base JSON-RPC `TransactionFilter` type and
having to respond that certain filters are unsupported, define our own
compatible type which only contains the variants we do support. This way
we will still return an invalid params error if someone tries to use an
unsupported filter, but the filters we advertise in the schema will
match what we can return a non-error response for.

## Test plan

Existing tests.