Consensus Data Quarantining
## Description 

- All consensus output is held in memory until all checkpoints built
from that output have been committed. This ensures that the epoch db is
not corrupted in the event of a fork.
- Consensus therefor becomes the primary replay log for all crash
recovery.
- Additionally this means there are several tables that we no longer
need to write to, and their contents can live in memory.

## Test plan 

Extensive testing with both simtest and antithesis. In particular the
upgrade flow (from non-DQ to DQ build) has been tested extensively with
antithesis.