fix!: Remove double indirection in `MoveFieldLayout`
## Description 

The `MoveFieldLayout` struct `fields` was wrapped in a box as part of
https://github.com/GeunhwaJeong/haneul/pull/19310, which states in its
reasoning:

> Boxes fields in the MoveTypeLayout to make the enum size smaller for
type layouts.

However, this field being boxed does not affect the enums which this was
referring to as it is also boxed in the variants. This means that the
variant size of those enums is always the size of a pointer. Thus, this
additional box around the vec serves only to create an additional
indirection.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: