[nit] Minor improvements to Object::new_coin
## Description 

This PR improves a few styles for clarity:
1. new_coin API should take the type_tag of the coin directly, instead
of allowing arbitrary move types.
2. the implementation of new_coin API should use Coin instead of
GasCoin, since it's not guaranteed to be a GasCoin.
3. Some other type improvements

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------