rpc-alt: Macro for building diesel `SqlLiteral`s
## Description

Introduce a macro that allows for building a `SqlLiteral` using a
format-string-like syntax.

`SqlLiteral` should allow us to keep using `diesel` and not have to
worry about how to handle scenarios where we need to introduce snippets
of raw SQL -- it makes it possible to do so while still taking advantage
of strong typing on the boundary and diesel's support for escaping
values.

However, it has poor ergonomics, requiring lots of boilerplate (chains
of function calls and type annotations) which break up and hide the
query being built. The newly introduced macro allows for building the
same structure using a format string. See diff for docs and examples of
usage.

Some features that were not included in the initial implementation, but
could be added later include:

- inlining binds (similar to inlining interpolated values in format
strings:
  - `sql!("foo = {bar:BigInt}")` instead of
  - `sql!("foo = {BigInt}", bar)`
- named binds (again, similar to the same feature in `format!` strings):
  - `sql!("foo = {bar:BigInt}", bar = 1 + 2)`.
- support for normal format strings, which in particular would allow us
to interpolate column and table names into sql snippets.

## Test plan

The macro comes with extensive unit tests, and the usages are covered by
existing tests:

```
cargo nextest run            \
  -p haneul-sql-macro           \
  -p haneul-indexer-alt         \
  -p haneul-indexer-alt-jsonrpc \
  -p haneul-indexer-alt-e2e-tests
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: