refactor(rpc-alt): pass configs via Context
## Description

Gather all the configs together and expose them through the `Context`,
rather than handing out the configs needed to each module as a field.

The motivation for this came from working on Display -- part of handling
object responses, which shows up across multiple RPC modules (objects,
query objects, dynamic fields), but deeply nested in the call stack and
requires some configuration.

With the old method of propagating configs, we would need to make the
display config available to all the modules that handle object responses
and thread it through the function call stack, even if e.g. dynamic
fields use object responses but never ask for Display.

With this new system, we piggy-back on the fact that the `Context`
parameter is already threaded into all these functions.

## Test plan

```
haneul$ cargo nextest run         \
  -p haneul-indexer-alt-jsonrpc   \
  -p haneul-indexer-alt-e2e-tests
```