[Consensus] Optimistically accept certified commits votes.
## Description 

The PR is utilising further the `CertifiedCommit` struct to piggy back
the votes that certify a commit. It is also extending the `CommitSyncer`
so it does utilise now internally the `CertifiedCommit` and better
manage the fetched blocks etc. The received votes are attempted to be
optimistically accepted via the normal path of the block manager. Pay
attention that in order to avoid kicking off the scheduled synchronizer
due to missing dependencies, the earlier mechanism that was introduced
which would allow missing ancestors to be fetched up to above a certain
threshold from the highest accepted round, will now need to get
disabled. This was added due to this
https://github.com/GeunhwaJeong/haneul/pull/19763 . But now that we commit
the certified commits directly, this won't be an issue.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: