Docs/stablecoin/usdc usage
## Description 

Updates the stablecoin section of the docs to include example usage of
the USDC stable coin in Move and PTBs, as well as provide some
additional context on the coin standard.

I removed the old example since a lot of it was related to the frontend
setup rather than the actual use of USDC. I made sure to keep the new
example focused on USDC assuming the reader already knows how to use the
coin standard and set up PTBs.

## Test plan 

---

## Release notes

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------