rpc-alt: Introduce IndexerCluster
## Description

This type combines the `Indexer` and `MetricsService` in one combined
cluster. It is also supported by a grouping of command-line args and
some re-exported modules to minimise the number of dependencies an
indexer developed outside the mono-repo needs to take.

## Test plan

New test that the cluster starts, runs, stops and updates the DB and
metrics:

```
haneul$ cargo nextest run -p haneul-indexer-alt-framework \
  -- test_indexer_cluster
```